# src/main/resources/swagger/customer-service.yml
openapi: 3.0.0
info:
  title: Client API
  version: 1.0.0
  description: API for managing bank clients. BootCamp Microservices 49.
  contact:
    name: No0dl3 Support
    url: https://github.com/byNo0dle
servers:
  - url: http://localhost:9003/api
    description: Client Service
paths:
  /client:
    get:
      tags:
        - Client
      summary: Find all Clients
      operationId: findAllClient
      responses:
        '200':
          description: List of Clients retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
        '400':
          description: Bad Request invalid syntax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientError'
        '404':
          description: Not Found Client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientError'
    post:
      tags:
        - Client
      summary: Add a new Client to the bank
      description: Add a new Client to the bank
      operationId: saveClient
      requestBody:
        description: Create a new Client in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
        required: true
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          description: Bad Request invalid syntax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientError'
        '404':
          description: Not Found Client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientError'
    put:
      tags:
        - Client
      summary: Update an existing Client
      description: Update an existing Client by Id
      operationId: updateClient
      requestBody:
        description: Update an existent Client in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          description: Bad Request invalid syntax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientError'
        '404':
          description: Not Found Client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientError'
  /client/{id}:
    get:
      tags:
        - Client
      summary: Find Client by ID
      description: Returns a single Client
      operationId: findByIdClient
      parameters:
        - name: id
          in: path
          description: ID of Client to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          description: Bad Request invalid syntax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientError'
        '404':
          description: Not Found Client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientError'
    delete:
      tags:
        - Client
      summary: Deletes a Client
      description: Delete a Client
      operationId: deleteClient
      parameters:
        - name: id
          in: path
          description: Client id to delete
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Not Content client
        '404':
          description: Not Found Client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientError'
  /client/numberDocumentIdentity/{numberDocumentIdentity}:
    get:
      tags:
        - Client
      summary: Search Client by numberDocumentIdentity
      description: Search Client by numberDocumentIdentity
      operationId: findByNumberDocumentIdentityClient
      parameters:
        - name: numberDocumentIdentity
          in: path
          description: This looks for a Client by numberDocumentIdentity
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          description: Bad Request invalid syntax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientError'
        '404':
          description: Not Found Client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientError'

components:
  schemas:
    Client:
      type: object
      properties:
        id:
          type: string
          example: 65xd2dx0x7d11d2x20827638
        firstName:
          type: string
          example: Piero Rogelio
        lastName:
          type: string
          example: Ahuanari Sangama
        emailAddress:
          type: string
          example: pieroahuanarisangama@gmail.com
        phoneNumber:
          type: string
          example: 981675432
        gender:
          type: string
          example: Masculino
        numberDocumentIdentity:
          type: string
          example: 70384520
        status:
          type: string
          example: ACTIVO
        isActive:
          type: boolean
          example:
            - true
            - false
        direction:
          $ref: '#/components/schemas/Direction'
        profileCustomer:
          $ref: '#/components/schemas/ProfileCustomer'
        typeCustomer:
          $ref: '#/components/schemas/TypeCustomer'
        identityDocumentType:
          $ref: '#/components/schemas/IdentityDocumentType'
    Direction:
      properties:
        type:
          type: string
          example: Av.
        name:
          type: string
          example: Pachitea
        number:
          type: string
          example: # 333
        province:
          type: string
          example: Coronel Portillo
        district:
          type: string
          example: Calleria
        department:
          type: string
          example: Ucayali
    ProfileCustomer:
      properties:
        code:
          type: string
          example: 1
        description:
          type: string
          example: Vip
        status:
          type: boolean
          example:
            - true
            - false
    TypeCustomer:
      properties:
        code:
          type: string
          example: 1
        description:
          type: string
          example: Personal
        isActive:
          type: boolean
          example:
            - true
            - false
    IdentityDocumentType:
      properties:
        code:
          type: string
          example: 1
        description:
          type: string
          example: Dni
        isActive:
          type: boolean
          example:
            - true
            - false
    ClientError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message